+++ moeflow_companion/mit_workflow/__init__.py
@@ -1,9 +1,12 @@
 from .process_file import process_files
 from ._const import is_cuda_avaiable, create_unique_dir
+from ._model import FileBatchProcessResult, FileProcessResult
 
 __all__ = [
     # "process_file",
     "process_files",
     "is_cuda_avaiable",
     "create_unique_dir",
+    "FileBatchProcessResult",
+    "FileProcessResult",
 ]
